конец, последняя версия раьоты

+++ DynamicArrays/DynamicArrays/ChangeArray.cpp
@@ -10,7 +10,7 @@ void ChangeArray(double* array, int32_t size, double number)
 	std::cout << "sorted by your number\n";
 	for (size_t i{}; i < size; ++i) 
 	{
-		for (int32_t j{}; j < size - i - 1; ++j)
+		for (size_t j{}; j < size - i - 1; ++j)
 		{
 			if (array[j] >= number && array[j + 1] < number)
 			{
