+++ DynamicArrays/DynamicArrays/SortArraySecond.cpp
@@ -2,9 +2,9 @@
 #include "Base.h"
 void SortArray(double* a, int32_t size)
 {
-	for (int32_t i{}; i < size - 1; ++i)
+	for (size_t i{}; i < size - 1; ++i)
 	{
-		for (int32_t j{}; j < size - i - 1; ++j)
+		for (size_t j{}; j < size - i - 1; ++j)
 		{
 			if (a[j] > a[j + 1] )
 			{
